Add CLI messages
