--- json-server.json
@@ -1,8 +1,0 @@
-{
-  "port": 5000,
-  "static": "dist",
-  "host": "localhost",
-  "watch": true,
-  "routes": "",
-  "delay": 0
-}
