--- middleware.js
@@ -1,5 +1,0 @@
-// hello.js
-module.exports = (req, res, next) => {
-  res.header('X-Hello', 'World');
-  next();
-};
