+++ package.json
@@ -27,6 +27,7 @@
     "chalk": "^4.1.0",
     "chokidar": "^3.5.1",
     "connect-pause": "^0.1.0",
+    "consola": "^2.15.3",
     "copyfiles": "^2.2.0",
     "core-js": "^3.6.5",
     "faker": "^4.1.0",
